requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.3.0
async-timeout==4.0.3
basho-erlastic==2.1.1
black==24.4.2
bson==0.5.10
cassandra-driver==3.29.1
certifi==2024.2.2
click==8.1.7
datamodel-code-generator==0.25.6
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastapi-redis-cache==0.2.5
gears-cli==1.2.0
genson==1.2.2
geomet==0.2.1.post1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inflect==5.6.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.4.0
mypy-extensions==1.0.0
odmantic==1.0.2
orjson==3.10.2
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pika==1.3.2
pika-stubs==0.1.3
platformdirs==4.2.1
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python3-protobuf==2.5.0
PyYAML==6.0.1
redis==4.6.0
riak==2.7.0.post28
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
